READ COMMIT MESSAGE BEFORE UNDOING
Added support for editing the term translation.
Changed term translation widget (tried TextInput but seems not to work well, so using TextArea with 1 row).
Added definition column.
Changed columns width (TODO this impacts on other parts of Pootle. Changing column classes might not to be an option since they are used in Javascript).
See if commented stuff in term_edit template is necessary (at least translator comments seem to be used to save other alternative translations for the term).